requirements.txt
numpy==2.4.2
opencv-python==4.13.0.92
pillow==12.1.1
pyserial==3.5